One task down out of four
There are still three remaining problems. the other two feature have not yet be programmed yet. the thing also won't disappear. after those are finish i will get more people to test out the code and will make the css of the context menu better and remove some things that are not need(if have permission).